Versão 2.5 - Implementação RF009, RF010 e RF011   - RF009 -> Consultar lista de questões abertas   - RF010 -> Consultar lista de questões resolvidas   - RF011 -> Consultar lista de questões por linguagem de programação

+++ src/java/persistence/linguagem_programacao/LinguagemProgramacaoDAO.java
@@ -17,7 +17,7 @@ public interface LinguagemProgramacaoDAO
 {
     public void cadastrarLinguagemProgramacao(LinguagemProgramacao p); //CREATE
     
-    public LinguagemProgramacao getLinguagemProgramacao(Long id); //READ
+    public LinguagemProgramacao getLinguagemProgramacao(Integer id); //READ
     
     public void atualizarLinguagemProgramacao(LinguagemProgramacao p); //UPDATE
